Update some routing groups
