+++ backend/app/api/routers.py
@@ -2,19 +2,19 @@
 # -*- coding: utf-8 -*-
 from fastapi import APIRouter
 
-from backend.app.core.conf import settings
+from backend.app.api.v1.api import router as api_router
 from backend.app.api.v1.auth import router as auth_router
-from backend.app.api.v1.user import router as user_router
 from backend.app.api.v1.casbin import router as casbin_router
 from backend.app.api.v1.dept import router as dept_router
-from backend.app.api.v1.role import router as role_router
-from backend.app.api.v1.menu import router as menu_router
-from backend.app.api.v1.api import router as api_router
-from backend.app.api.v1.login_log import router as login_log_router
-from backend.app.api.v1.opera_log import router as opera_log_router
-from backend.app.api.v1.dict_type import router as dict_type_router
 from backend.app.api.v1.dict_data import router as dict_data_router
+from backend.app.api.v1.dict_type import router as dict_type_router
+from backend.app.api.v1.log import router as log_router
+from backend.app.api.v1.menu import router as menu_router
 from backend.app.api.v1.mixed import router as mixed_router
+from backend.app.api.v1.monitor import router as monitor_router
+from backend.app.api.v1.role import router as role_router
+from backend.app.api.v1.user import router as user_router
+from backend.app.core.conf import settings
 
 v1 = APIRouter(prefix=settings.API_V1_STR)
 
@@ -27,6 +27,6 @@ v1.include_router(menu_router, prefix='/menus', tags=['菜单管理'])
 v1.include_router(api_router, prefix='/apis', tags=['API管理'])
 v1.include_router(dict_type_router, prefix='/dict-types', tags=['字典类型管理'])
 v1.include_router(dict_data_router, prefix='/dict-datas', tags=['字典数据管理'])
-v1.include_router(login_log_router, prefix='/login-logs', tags=['登录日志管理'])
-v1.include_router(opera_log_router, prefix='/opera-logs', tags=['操作日志管理'])
+v1.include_router(log_router, prefix='/logs', tags=['日志管理'])
+v1.include_router(monitor_router, prefix='/monitors', tags=['监控管理'])
 v1.include_router(mixed_router, prefix='/mixes', tags=['杂项'])
