+++ backend/app/api/v1/log/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+from fastapi import APIRouter
+from backend.app.api.v1.log.login_log import router as login_log
+from backend.app.api.v1.log.opera_log import router as opera_log
+
+router = APIRouter()
+
+router.include_router(login_log)
+router.include_router(opera_log)
